Modify #11 - AnswerDao 인터페이스
의존관계에 대한 결합을 줄이기 위해 인터페이스 추출
